adding triggers

+++ deployment/notifications.py
@@ -1,24 +1,41 @@
 import boto3
+import time
+from deployment.falsk_app.process import trigger_main_process
+from src.constants.constants import *
 
-AWS_ACCESS_KEY = 'AKIAIR6DQWSOVXCJAXPA'
-AWS_SECRET_ACCESS_KEY = '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'
+ACCESS_ID = USER1_ACCESS_ID
+ACCESS_KEY = USER1_ACCESS_KEY
 
-SQS_QUEUE_NAME = 'fyp_company_dataq'
 
-sqs_resource_connection = boto3.resource(
-    'sqs',
-    aws_access_key_id=AWS_ACCESS_KEY,
-    aws_secret_access_key=AWS_SECRET_ACCESS_KEY,
-    region_name='ap-southeast-1'
-)
+def alert_for_sqs_notifications():
+    sqs_resource_connection = boto3.resource(
+        'sqs',
+        aws_access_key_id=ACCESS_ID,
+        aws_secret_access_key=ACCESS_KEY,
+        region_name='ap-southeast-1'
+    )
 
-queue = sqs_resource_connection.Queue('https://sqs.ap-southeast-1.amazonaws.com/190366723864/fyp_company_dataq')
+    queue = sqs_resource_connection.Queue('https://sqs.ap-southeast-1.amazonaws.com/190366723864/fyp_company_dataq')
 
-# queue = sqs_resource_connection.get_queue_by_name(QueueName=SQS_QUEUE_NAME)
+    # queue = sqs_resource_connection.get_queue_by_name(QueueName=SQS_QUEUE_NAME)
+    while True:
+        messages = queue.receive_messages(MaxNumberOfMessages=1, WaitTimeSeconds=2)
+        print('...........')
+        if messages:
+            for message in messages:
+                print(message.body)
+                message.delete()  # delete the message from the queue
+                trigger_main_process()
+            return True
+        else:
+            print('no msgs')
+            time.sleep(2)
+            # return False
 
-messages = queue.receive_messages(MaxNumberOfMessages=5, WaitTimeSeconds=10)
-for message in messages:
-    print(message.body)
-    message.delete()  # delete the message from the queue
 
-response = queue.send_message(MessageBody='Hello by AWS SDK for Python!')
+    # for message in messages:
+    #     print(message.body)
+    #     message.delete()  # delete the message from the queue
+    #
+    # response = queue.send_message(MessageBody='Hello by AWS SDK for Python!')
+# alert_for_sqs_notifications()
